refactor: add MultiSelectOptionItem
