+++ src/app/Components/ArtworkFilter/Filters/MultiSelectOption.tsx
@@ -3,16 +3,10 @@ import { StackNavigationProp } from "@react-navigation/stack"
 import { FilterData } from "app/Components/ArtworkFilter/ArtworkFilterHelpers"
 import { ArtworkFilterBackHeader } from "app/Components/ArtworkFilter/components/ArtworkFilterBackHeader"
 import { SearchInput } from "app/Components/SearchInput"
-import { TouchableRow } from "app/Components/TouchableRow"
-import { Box, Check, CHECK_SIZE, Flex, Text, useSpace } from "palette"
-import React, { memo, useCallback, useState } from "react"
+import { Flex, Text } from "palette"
+import React, { useCallback, useState } from "react"
 import { FlatList, ScrollView } from "react-native"
-import { useScreenDimensions } from "shared/hooks"
-import styled from "styled-components/native"
-
-const OPTIONS_MARGIN_LEFT = 0.5
-const OPTION_PADDING = 15
-const OPTION_HEIGHT = 50
+import { MULTI_SELECT_OPTION_ITEM_HEIGHT, MultiSelectOptionItem } from "./MultiSelectOptionItem"
 
 interface MultiSelectOptionScreenProps {
   navigation: StackNavigationProp<ParamListBase>
@@ -140,8 +134,8 @@ export const MultiSelectOptionScreen: React.FC<MultiSelectOptionScreenProps> = (
             renderItem={renderItem}
             windowSize={11}
             getItemLayout={(_, index) => ({
-              length: OPTION_HEIGHT,
-              offset: index * OPTION_HEIGHT,
+              length: MULTI_SELECT_OPTION_ITEM_HEIGHT,
+              offset: index * MULTI_SELECT_OPTION_ITEM_HEIGHT,
               index,
             })}
           />
@@ -150,57 +144,3 @@ export const MultiSelectOptionScreen: React.FC<MultiSelectOptionScreenProps> = (
     </Flex>
   )
 }
-
-interface MultiSelectOptionItemProps {
-  item: FilterData
-  selected: boolean
-  disabled: boolean
-  onPress: (item: FilterData) => void
-}
-
-const areEqual = (prevProps: MultiSelectOptionItemProps, nextProps: MultiSelectOptionItemProps) => {
-  return (
-    prevProps.selected === nextProps.selected &&
-    prevProps.disabled === nextProps.disabled &&
-    prevProps.item.displayText === nextProps.item.displayText &&
-    prevProps.onPress === nextProps.onPress
-  )
-}
-
-const MultiSelectOptionItem: React.FC<MultiSelectOptionItemProps> = memo(
-  ({ item, disabled, selected, onPress }) => {
-    const space = useSpace()
-    const { width } = useScreenDimensions()
-    const optionTextMaxWidth = width - OPTION_PADDING * 3 - space(OPTIONS_MARGIN_LEFT) - CHECK_SIZE
-
-    return (
-      <Box ml={OPTIONS_MARGIN_LEFT} height={OPTION_HEIGHT}>
-        <TouchableRow
-          onPress={() => onPress(item)}
-          disabled={disabled}
-          testID="multi-select-option-button"
-          accessibilityState={{ checked: selected }}
-        >
-          <OptionListItem>
-            <Box maxWidth={optionTextMaxWidth}>
-              <Text variant="xs" color="black100">
-                {item.displayText}
-              </Text>
-            </Box>
-
-            <Check selected={selected} disabled={disabled} />
-          </OptionListItem>
-        </TouchableRow>
-      </Box>
-    )
-  },
-  areEqual
-)
-
-export const OptionListItem = styled(Flex)`
-  flex-direction: row;
-  justify-content: space-between;
-  flex-grow: 1;
-  align-items: flex-start;
-  padding: ${OPTION_PADDING}px;
-`
